Refactor Tap form into its own component
