+++ web/app/js/components/Tap.jsx
@@ -5,21 +5,10 @@ import PropTypes from 'prop-types';
 import { publicAddressToString } from './util/Utils.js';
 import React from 'react';
 import TapEventTable from './TapEventTable.jsx';
+import TapQueryForm from './TapQueryForm.jsx';
 import { withContext } from './util/AppContext.jsx';
-import {
-  AutoComplete,
-  Button,
-  Col,
-  Form,
-  Icon,
-  Input,
-  Row,
-  Select
-} from 'antd';
 import './../../css/tap.css';
 
-const colSpan = 5;
-const rowGutter = 16;
 const httpMethods = ["GET", "HEAD", "POST", "PUT", "DELETE", "CONNECT", "OPTIONS", "TRACE", "PATCH"];
 const defaultMaxRps = 1.0;
 const maxNumFilterOptions = 12;
@@ -53,17 +42,9 @@ class Tap extends React.Component {
         authority: "",
         maxRps: defaultMaxRps
       },
-      autocomplete: {
-        namespace: [],
-        resource: [],
-        toNamespace: [],
-        toResource: [],
-        authority: []
-      },
       maxLinesToDisplay: 40,
       awaitingWebSocketConnection: false,
       tapRequestInProgress: false,
-      showAdvancedForm: false,
       pollingInterval: 10000,
       pendingRequests: false
     };
@@ -158,10 +139,6 @@ class Tap extends React.Component {
       authoritiesByNs,
       resourcesByNs
     };
-  }
-
-  getResourceList(resourcesByNs, ns) {
-    return resourcesByNs[ns] || _.uniq(_.flatten(_.values(resourcesByNs)));
   }
 
   parseTapResult = data => {
@@ -311,12 +288,6 @@ class Tap extends React.Component {
     });
   }
 
-  toggleAdvancedForm = show => {
-    this.setState({
-      showAdvancedForm: show
-    });
-  }
-
   handleTapStart = e => {
     e.preventDefault();
     this.startTapSteaming();
@@ -324,44 +295,6 @@ class Tap extends React.Component {
 
   handleTapStop = () => {
     this.ws.close(1000);
-  }
-
-  handleFormChange = (name, scopeResource, shouldScopeAuthority) => {
-    let state = {
-      query: this.state.query,
-      autocomplete: this.state.autocomplete
-    };
-
-    return formVal => {
-      state.query[name] = formVal;
-      if (!_.isNil(scopeResource)) {
-        // scope the available typeahead resources to the selected namespace
-        state.autocomplete[scopeResource] = this.state.resourcesByNs[formVal];
-      }
-      if (shouldScopeAuthority) {
-        state.autocomplete.authority = this.state.authoritiesByNs[formVal];
-      }
-
-      this.setState(state);
-    };
-  }
-
-  handleFormEvent = name => {
-    let state = {
-      query: this.state.query
-    };
-
-    return event => {
-      state.query[name] = event.target.value;
-      this.setState(state);
-    };
-  }
-
-  autoCompleteData = name => {
-    return _(this.state.autocomplete[name])
-      .filter(d => d.indexOf(this.state.query[name]) !== -1)
-      .sortBy()
-      .value();
   }
 
   loadFromServer() {
@@ -377,21 +310,10 @@ class Tap extends React.Component {
     this.serverPromise = Promise.all(this.api.getCurrentPromises())
       .then(rsp => {
         let { resourcesByNs, authoritiesByNs } = this.getResourcesByNs(rsp);
-        let namespaces = _.sortBy(_.keys(resourcesByNs));
-        let resourceNames  = this.getResourceList(resourcesByNs, this.state.query.namespace);
-        let toResourceNames = this.getResourceList(resourcesByNs, this.state.query.toNamespace);
-        let authorities = this.getResourceList(authoritiesByNs, this.state.query.namespace);
 
         this.setState({
           resourcesByNs,
           authoritiesByNs,
-          autocomplete: {
-            namespace: namespaces,
-            resource: resourceNames,
-            toNamespace: namespaces,
-            toResource: toResourceNames,
-            authority: authorities
-          },
           pendingRequests: false
         });
       })
@@ -407,156 +329,6 @@ class Tap extends React.Component {
       pendingRequests: false,
       error: e
     });
-  }
-
-  renderTapForm = () => {
-    return (
-      <Form className="tap-form">
-        <Row gutter={rowGutter}>
-          <Col span={colSpan}>
-            <Form.Item>
-              <Select
-                showSearch
-                allowClear
-                placeholder="Namespace"
-                optionFilterProp="children"
-                onChange={this.handleFormChange("namespace", "resource", true)}>
-                {
-                  _.map(this.state.autocomplete.namespace, (n, i) => (
-                    <Select.Option key={`ns-dr-${i}`} value={n}>{n}</Select.Option>
-                  ))
-                }
-              </Select>
-            </Form.Item>
-          </Col>
-
-          <Col span={colSpan}>
-            <Form.Item>
-              <AutoComplete
-                dataSource={this.autoCompleteData("resource")}
-                onSelect={this.handleFormChange("resource")}
-                onSearch={this.handleFormChange("resource")}
-                placeholder="Resource" />
-            </Form.Item>
-          </Col>
-
-          <Col span={colSpan}>
-            <Form.Item>
-              {
-                this.state.tapRequestInProgress ?
-                  <Button type="primary" className="tap-stop" onClick={this.handleTapStop}>Stop</Button> :
-                  <Button type="primary" className="tap-start" onClick={this.handleTapStart}>Start</Button>
-              }
-              {
-                this.state.awaitingWebSocketConnection ?
-                  <Icon type="loading" style={{ paddingLeft: rowGutter, fontSize: 20, color: '#08c' }} /> : null
-              }
-            </Form.Item>
-          </Col>
-        </Row>
-
-        <Button
-          className="tap-form-toggle"
-          onClick={() => this.toggleAdvancedForm(!this.state.showAdvancedForm)}>
-          { this.state.showAdvancedForm ?
-            "Hide filters" : "Show more request filters" } <Icon type={this.state.showAdvancedForm ? 'up' : 'down'} />
-        </Button>
-
-        { !this.state.showAdvancedForm ? null : this.renderAdvancedTapForm() }
-      </Form>
-    );
-  }
-
-  renderAdvancedTapForm = () => {
-    return (
-      <React.Fragment>
-        <Row gutter={rowGutter}>
-          <Col span={colSpan}>
-            <Form.Item>
-              <Select
-                showSearch
-                allowClear
-                placeholder="To Namespace"
-                optionFilterProp="children"
-                onChange={this.handleFormChange("toNamespace", "toResource")}>
-                {
-                  _.map(this.state.autocomplete.toNamespace, (n, i) => (
-                    <Select.Option key={`ns-dr-${i}`} value={n}>{n}</Select.Option>
-                  ))
-                }
-              </Select>
-            </Form.Item>
-          </Col>
-
-          <Col span={colSpan}>
-            <Form.Item>
-              <AutoComplete
-                dataSource={this.autoCompleteData("toResource")}
-                onSelect={this.handleFormChange("toResource")}
-                onSearch={this.handleFormChange("toResource")}
-                placeholder="To Resource" />
-            </Form.Item>
-          </Col>
-        </Row>
-
-        <Row gutter={rowGutter}>
-          <Col span={2 * colSpan}>
-            <Form.Item
-              extra="Display requests with this :authority">
-
-              <AutoComplete
-                dataSource={this.autoCompleteData("authority")}
-                onSelect={this.handleFormChange("authority")}
-                onSearch={this.handleFormChange("authority")}
-                placeholder="Authority" />
-            </Form.Item>
-          </Col>
-
-          <Col span={2 * colSpan}>
-            <Form.Item
-              extra="Display requests with paths that start with this prefix">
-              <Input placeholder="Path" onChange={this.handleFormEvent("path")} />
-            </Form.Item>
-          </Col>
-
-        </Row>
-
-        <Row gutter={rowGutter}>
-          <Col span={colSpan}>
-            <Form.Item
-              extra="Display requests with this scheme">
-              <Input placeholder="Scheme" onChange={this.handleFormEvent("scheme")} />
-            </Form.Item>
-          </Col>
-
-          <Col span={colSpan}>
-            <Form.Item
-              extra="Maximum requests per second to tap">
-              <Input
-                defaultValue={defaultMaxRps}
-                placeholder="Max RPS"
-                onChange={this.handleFormEvent("maxRps")} />
-            </Form.Item>
-          </Col>
-
-          <Col span={colSpan}>
-            <Form.Item
-              extra="Display requests with this HTTP method">
-              <Select
-                allowClear
-                placeholder="HTTP method"
-                onChange={this.handleFormChange("method")}>
-                {
-                  _.map(httpMethods, m =>
-                    <Select.Option key={`method-select-${m}`} value={m}>{m}</Select.Option>
-                  )
-                }
-              </Select>
-            </Form.Item>
-          </Col>
-        </Row>
-      </React.Fragment>
-    );
   }
 
   renderCurrentQuery = () => {
@@ -593,7 +365,14 @@ class Tap extends React.Component {
         <ErrorBanner message={this.state.error} onHideMessage={() => this.setState({ error: null })} />}
 
         <PageHeader header="Tap" />
-        {this.renderTapForm()}
+        <TapQueryForm
+          tapRequestInProgress={this.state.tapRequestInProgress}
+          awaitingWebSocketConnection={this.state.awaitingWebSocketConnection}
+          handleTapStart={this.handleTapStart}
+          handleTapStop={this.handleTapStop}
+          resourcesByNs={this.state.resourcesByNs}
+          authoritiesByNs={this.state.authoritiesByNs}
+          query={this.state.query} />
         {this.renderCurrentQuery()}
 
         <TapEventTable
